Renaming command sections, and restructured around device watch commands
