+++ trueseeing/app/cmd/android/exploit.py
@@ -7,37 +7,40 @@ from trueseeing.core.model.cmd import CommandMixin
 from trueseeing.core.ui import ui
 
 if TYPE_CHECKING:
-  from typing import Any, Optional, Dict, Mapping
+  from typing import Any, Optional, Dict, Mapping, List, Iterator, Tuple
   from trueseeing.api import CommandHelper, Command, CommandMap, OptionMap
   from trueseeing.core.android.context import APKContext
 
-class ExploitCommand(CommandMixin):
+class EngageCommand(CommandMixin):
   def __init__(self, helper: CommandHelper) -> None:
     self._helper = helper
 
   @staticmethod
   def create(helper: CommandHelper) -> Command:
-    return ExploitCommand(helper)
+    return EngageCommand(helper)
 
   def get_commands(self) -> CommandMap:
     return {
-      'xq':dict(e=self._exploit_discard, n='xq', d='exploit: discard changes'),
-      'xx':dict(e=self._exploit_apply, n='xx[!]', d='exploit: apply and rebuild apk'),
-      'xx!':dict(e=self._exploit_apply),
-      'xf':dict(e=self._exploit_inject_frida, n='xf[!] [config]', d='exploit; inject frida gadget'),
-      'xf!':dict(e=self._exploit_inject_frida),
-      'xfs':dict(e=self._exploit_inject_frida_scriptdir, n='xfs[!] [path]', d='exploit; inject frida gadget in script dir mode'),
-      'xfs!':dict(e=self._exploit_inject_frida_scriptdir),
-      'xu':dict(e=self._exploit_disable_pinning, n='xu', d='exploit: disable SSL/TLS pinning'),
-      'xd':dict(e=self._exploit_enable_debug, n='xd', d='exploit: make debuggable'),
-      'xb':dict(e=self._exploit_enable_backup, n='xb', d='exploit: make backupable'),
-      'xt':dict(e=self._exploit_patch_target_api_level, n='xt[!] <api level>', d='exploit: patch target api level'),
-      'xt!':dict(e=self._exploit_patch_target_api_level),
-      'xp':dict(e=self._exploit_device_list_packages, n='xp', d='device: list installed packages'),
-      'xco':dict(e=self._exploit_device_copyout, n='xco[!] package [data.tar]', d='device: copy-out package data'),
-      'xco!':dict(e=self._exploit_device_copyout),
-      'xci':dict(e=self._exploit_device_copyin, n='xci[!] package [data.tar]', d='device: copy-in package data'),
-      'xci!':dict(e=self._exploit_device_copyin),
+      'xq':dict(e=self._engage_discard, n='xq', d='engage: discard changes'),
+      'xx':dict(e=self._engage_apply, n='xx[!]', d='engage: apply and rebuild apk'),
+      'xx!':dict(e=self._engage_apply),
+      'xf':dict(e=self._engage_inject_frida, n='xf[!] [config]', d='engage; inject frida gadget'),
+      'xf!':dict(e=self._engage_inject_frida),
+      'xfs':dict(e=self._engage_inject_frida_scriptdir, n='xfs[!] [path]', d='engage; inject frida gadget in script dir mode'),
+      'xfs!':dict(e=self._engage_inject_frida_scriptdir),
+      'xu':dict(e=self._engage_disable_pinning, n='xu', d='engage: disable SSL/TLS pinning'),
+      'xd':dict(e=self._engage_enable_debug, n='xd', d='engage: make debuggable'),
+      'xb':dict(e=self._engage_enable_backup, n='xb', d='engage: make backupable'),
+      'xt':dict(e=self._engage_patch_target_api_level, n='xt[!] <api level>', d='engage: patch target api level'),
+      'xt!':dict(e=self._engage_patch_target_api_level),
+      'xco':dict(e=self._engage_device_copyout, n='xco[!] package [data.tar]', d='engage: copy-out package data'),
+      'xco!':dict(e=self._engage_device_copyout),
+      'xci':dict(e=self._engage_device_copyin, n='xci[!] package [data.tar]', d='engage: copy-in package data'),
+      'xci!':dict(e=self._engage_device_copyin),
+      'xz':dict(e=self._engage_fuzz_intent, n='xz[!] "am-cmdline-template" [output.txt]', d='engage: fuzz intent'),
+      'xz!':dict(e=self._engage_fuzz_intent),
+      'xzr':dict(e=self._engage_fuzz_command, n='xzr[!] "cmdline-template" [output.txt]', d='engage: fuzz cmdline'),
+      'xzr!':dict(e=self._engage_fuzz_command),
     }
 
   def get_options(self) -> OptionMap:
@@ -45,7 +48,7 @@ class ExploitCommand(CommandMixin):
       'vers':dict(n='vers=X.Y.Z', d='specify frida-gadget version to use [xf,xfs]')
     }
 
-  async def _exploit_discard(self, args: deque[str]) -> None:
+  async def _engage_discard(self, args: deque[str]) -> None:
     apk = self._helper.require_target()
 
     _ = args.popleft()
@@ -63,7 +66,7 @@ class ExploitCommand(CommandMixin):
 
     ui.success('done ({t:.02f} sec.)'.format(t=(time.time() - at)))
 
-  async def _exploit_apply(self, args: deque[str]) -> None:
+  async def _engage_apply(self, args: deque[str]) -> None:
     apk = self._helper.require_target()
 
     cmd = args.popleft()
@@ -107,7 +110,7 @@ class ExploitCommand(CommandMixin):
 
     ui.success('done ({t:.02f} sec.)'.format(t=(time.time() - at)))
 
-  async def _exploit_disable_pinning(self, args: deque[str]) -> None:
+  async def _engage_disable_pinning(self, args: deque[str]) -> None:
     apk = self._helper.require_target()
 
     _ = args.popleft()
@@ -157,7 +160,7 @@ class ExploitCommand(CommandMixin):
 
     ui.success('done ({t:.02f} sec.)'.format(t=(time.time() - at)))
 
-  async def _exploit_enable_debug(self, args: deque[str]) -> None:
+  async def _engage_enable_debug(self, args: deque[str]) -> None:
     apk = self._helper.require_target()
 
     _ = args.popleft()
@@ -179,7 +182,7 @@ class ExploitCommand(CommandMixin):
 
     ui.success('done ({t:.02f} sec.)'.format(t=(time.time() - at)))
 
-  async def _exploit_enable_backup(self, args: deque[str]) -> None:
+  async def _engage_enable_backup(self, args: deque[str]) -> None:
     apk = self._helper.require_target()
 
     _ = args.popleft()
@@ -203,7 +206,7 @@ class ExploitCommand(CommandMixin):
 
     ui.success('done ({t:.02f} sec.)'.format(t=(time.time() - at)))
 
-  async def _exploit_patch_target_api_level(self, args: deque[str]) -> None:
+  async def _engage_patch_target_api_level(self, args: deque[str]) -> None:
     apk = self._helper.require_target()
 
     cmd = args.popleft()
@@ -239,7 +242,7 @@ class ExploitCommand(CommandMixin):
     ui.success('done ({t:.02f} sec.)'.format(t=(time.time() - at)))
 
   # XXX: long and ugly
-  async def _exploit_inject_frida(self, args: deque[str], script_dir_mode: bool = False) -> None:
+  async def _engage_inject_frida(self, args: deque[str], script_dir_mode: bool = False) -> None:
     configfn: Optional[str] = None
     config_override: Optional[str] = None
     dev_frida_dir: Optional[str] = None
@@ -419,8 +422,8 @@ class ExploitCommand(CommandMixin):
 
     ui.success('done ({t:.02f} sec.)'.format(t=(time.time() - at)))
 
-  async def _exploit_inject_frida_scriptdir(self, args: deque[str]) -> None:
-    await self._exploit_inject_frida(args, script_dir_mode=True)
+  async def _engage_inject_frida_scriptdir(self, args: deque[str]) -> None:
+    await self._engage_inject_frida(args, script_dir_mode=True)
 
   def _as_dalvik_classname(self, jn: str) -> str:
     return 'L{};'.format(jn.replace('.', '/'))
@@ -475,24 +478,7 @@ class ExploitCommand(CommandMixin):
       except ClientConnectionError:
         raise InvalidResponseError()
 
-  async def _exploit_device_list_packages(self, args: deque[str]) -> None:
-    _ = args.popleft()
-
-    ui.info('listing packages')
-
-    import time
-    import re
-    from trueseeing.core.android.device import AndroidDevice
-
-    at = time.time()
-    nr = 0
-    for m in re.finditer(r'^package:(.*)', await AndroidDevice().invoke_adb('shell pm list package'), re.MULTILINE):
-      p = m.group(1)
-      ui.info(p)
-      nr += 1
-    ui.success('done, {nr} packages found ({t:.02f} sec.)'.format(nr=nr, t=(time.time() - at)))
-
-  async def _exploit_device_copyout(self, args: deque[str]) -> None:
+  async def _engage_device_copyout(self, args: deque[str]) -> None:
     success: bool = False
 
     cmd = args.popleft()
@@ -585,7 +571,7 @@ class ExploitCommand(CommandMixin):
     else:
       ui.failure('copyout failed')
 
-  async def _exploit_device_copyin(self, args: deque[str]) -> None:
+  async def _engage_device_copyin(self, args: deque[str]) -> None:
     success: bool = False
 
     _ = args.popleft()
@@ -676,6 +662,104 @@ class ExploitCommand(CommandMixin):
       ui.success('done ({t:.02f} sec.)'.format(t=(time.time() - at)))
     else:
       ui.failure('copyin failed')
+
+  async def _engage_fuzz_command(self, args: deque[str], am: bool = False) -> None:
+    outfn: Optional[str] = None
+
+    cmd = args.popleft()
+
+    if not args:
+      if am:
+        ui.fatal('an "am" command line pattern required; try giving whatever you would to "adb shell am" (e.g. {} "start-activity .." ..)'.format(cmd))
+      else:
+        ui.fatal('command line pattern required; try giving you would to "adb shell"')
+
+    pat = args.popleft()
+    if am:
+      pat = f'am {pat}'
+
+    if args and not args[0].startswith('@'):
+      import os
+      outfn = args.popleft()
+      if os.path.exists(outfn) and not cmd.endswith('!'):
+        ui.fatal('outfile exists; force (!) to overwrite')
+
+    wordlist: Dict[str, List[str]] = dict()
+    for name, fn in self._helper.get_effective_options(self._helper.get_modifiers(args)).items():
+      if name.startswith('w'):
+        name = name[1:]
+        try:
+          with open(fn, 'r') as f:
+            wordlist[name] = [x.rstrip() for x in f]
+        except OSError as e:
+          ui.fatal(f'cannot open wordlist: {e}')
+
+    if not wordlist:
+      ui.fatal('need a wordlist (try @o:wNAME=FN)')
+
+    ui.info('wordlist built: {} words in {} keys ({})'.format(sum([len(v) for v in wordlist.values()]), len(wordlist), ','.join(wordlist.keys())))
+
+    def _expand(pat: str, wordlist: Mapping[str, List[str]]) -> Iterator[Tuple[int, int, str]]:
+      tries = min(len(v) for v in wordlist.values())
+      for nr in range(tries):
+        d = {k:v[nr] for k,v in wordlist.items()}
+        try:
+          yield nr, tries, pat.format(*[], **d)
+        except KeyError as e:
+          ui.fatal(f'unknown wordlist specified: {e}')
+
+    ui.info('starting fuzzing, opening log system-wide{}'.format(' [{}]'.format(outfn) if outfn else ''))
+
+    from trueseeing.core.android.device import AndroidDevice
+
+    dev = AndroidDevice()
+
+    async def _log(outfn: Optional[str]) -> None:
+      import sys
+      nr = 0
+
+      if not outfn:
+        f = sys.stdout.buffer
+      else:
+        f = open(outfn, 'wb')
+
+      try:
+        async for l in dev.invoke_adb_streaming('logcat -T1'):
+          f.write(l)
+          nr += 1
+          if outfn and nr % 256 == 0:
+            ui.info(' ... captured: {}')
+      finally:
+        if outfn:
+          f.close()
+
+    async def _fuzz(pat: str, wordlist: Mapping[str, List[str]]) -> None:
+      from asyncio import sleep
+      from subprocess import CalledProcessError
+      for nr, tries, t in _expand(pat, wordlist):
+        await sleep(.05)
+        prog = dict(nr=nr+1, max=tries, cmd=t)
+        try:
+          await dev.invoke_adb(f'shell {t}')
+          ui.info('[{nr}/{max}] {cmd}'.format(**prog))
+        except CalledProcessError as e:
+          ui.failure('[{nr}/{max}] {cmd}: failed: {code}'.format(code=e.returncode, **prog))
+
+    from asyncio import create_task, wait, FIRST_COMPLETED, ALL_COMPLETED
+    task_log = create_task(_log(outfn))
+    task_fuzz = create_task(_fuzz(pat, wordlist))
+
+    done, pending = await wait([task_log, task_fuzz], return_when=FIRST_COMPLETED)
+    for t in pending:
+      t.cancel()
+    done, _ = await wait([task_log, task_fuzz], return_when=ALL_COMPLETED)
+    for t in done:
+      exc = t.exception()
+      if exc:
+        ui.error('unhandled exception', exc=exc)
+
+  async def _engage_fuzz_intent(self, args: deque[str]) -> None:
+    await self._engage_fuzz_command(args, am=True)
 
   def _generate_tempfilename_for_device(self, dir: Optional[str] = None) -> str:
     import random
